Accounts added

+++ fridge_api/routes/mongoAPI.js
@@ -1,12 +1,19 @@
 const express = require("express");
 const foodModel = require("../models/foodModel");
+const requireAuth = require('../middleware/requireAuth');
+
 var router = express.Router();
 
+router.use(requireAuth);
+
 router.post("/add_food", async (request, response) => {
-    const food = new foodModel(request.body);
-    console.log(food);
+    // const food = new foodModel(request.body);
+    const { name, type, quantity } = request.body;
+
     try {
-      await food.save();
+      const user_id = request.user._id;
+      // await food.save();
+      const food = await foodModel.create({name, type, quantity, user_id})
       response.send(food);
     } catch (error) {
       response.status(500).send(error);
@@ -14,8 +21,8 @@ router.post("/add_food", async (request, response) => {
 });
 
 router.get("/foods", async (request, response) => {
-    const foods = await foodModel.find({});
-  
+    const user_id = request.user._id;
+    const foods = await foodModel.find({ user_id });
     try {
       response.send(foods);
     } catch (error) {
@@ -24,8 +31,9 @@ router.get("/foods", async (request, response) => {
 });
 
 router.get("/food", async (request, response) => {
+  const user_id = request.user._id;
   const name = request.query.param;
-  const food = await foodModel.find({ name });
+  const food = await foodModel.find({ name, user_id });
   console.log(food);
   try {
     response.send(food);
@@ -35,11 +43,12 @@ router.get("/food", async (request, response) => {
 });
 
 router.post("/update_food", async (request, response) => {
+    const user_id = request.user._id;
     const name = request.query.name;
     console.log(name);
     const quan = request.query.quan;
     console.log(quan);
-    const food = await foodModel.updateOne({ name: name }, { quantity: quan });
+    const food = await foodModel.updateOne({ name: name, user_id: user_id }, { quantity: quan });
     console.log(food);
     try {
       // await food.save();
@@ -50,9 +59,10 @@ router.post("/update_food", async (request, response) => {
 });
 
 router.delete("/delete_food", async (request, response) => {
+  const user_id = request.user._id;
   const name = request.query.name;
   console.log(name);
-  const food = await foodModel.deleteOne({ name: name });
+  const food = await foodModel.deleteOne({ name: name, user_id: user_id });
   try {
     console.log("Data Deleted");
     response.send(food);
